rename typing and use pipe syntax instead of Union
