+++ src/lr_reduction/mantid_utils.py
@@ -1,4 +1,4 @@
-from lr_reduction.typing import MantidWorkspace
+from lr_reduction.types import MantidWorkspace
 from lr_reduction.utils import workspace_handle
