+++ src/lr_reduction/types.py
@@ -1,0 +1,3 @@
+from mantid.api import Workspace
+
+MantidWorkspace = str | Workspace
