--- src/lr_reduction/typing.py
@@ -1,5 +1,0 @@
-from typing import Union
-
-from mantid.api import Workspace
-
-MantidWorkspace = Union[str, Workspace]
